+++ tests/testthat/test-tutorials_manynet.R
@@ -1,0 +1,7 @@
+test_that("manynet tutorials work", {
+  for(tute.dir in list.dirs(system.file("tutorials", package = "manynet"), 
+                            recursive = F)){
+    tute.file <- list.files(tute.dir, pattern = "*.Rmd", full.names = T)
+    expect_null(test_tutorials(tute.file))
+  }
+})